Move some functions into sdsc.tokens
* re_compile
* linenumber
* removepunctuation
* sanitizepunctuation
* findtagreplacement
* tokenizer
* counttokens
